describe what this is about
The name should be "related" to `CzechLight/gammarus` that we're using
internally for an on-device web dashboard. Taxonomically, Velia caprai
is a water bug (hence that connection to gammarus). It is not really a
"vodoměrka" (which would be obviously connected to monitoring), but a
"hladinatka". On the other hand, "velia" sounds about 666× times better
than Hydrometra stagnorum -- so "hladinatka" it is.

Change-Id: Iba564639bcaee302f6a37dd88775b49f5a81595f