Get rid of git submodules
We are not directly using any code that is prone to frequent API
changes (that's why we are pinning specific commits of the whole dep
stack in, say, netconf-cli). Let's try to make C++ dependency handling
much easier in this project:

- keep reusing the prebuilt binaries from CzechLight/dependencies
- do not pin a specific commit here, just use whatever Zuul throws at us

One therefore doesn't have to check out a specific commit anymore. In
case there's a Depends-on, we still pull stuff from a preceding job in
the pipeline.

Perhaps we could reuse this *major* simplification in those more fragile
leaf projects that do require specific versions of dependencies; we have
some multi-project integration through Zuul -- but that's for future
work.

Change-Id: I85549e689eef704f6f6eb5976e8972682c59b56e