Update patch set 8
Patch Set 8: Code-Review-1

(1 comment)

I would prefer this to be simplified. This patch proposes three classes, a source of information (AbstractInput), a reactor which processes the states (StateManager), and a connection that is used for value passing (StateManagerInput).

I don't fully understand the purpose of StateManagerInput, and I think one can get by without that class, just let StateManager query the AbstractInput classes directly (or let the AbstractInput push changes to the StateManager -- the point is to either clearly spell out the purpose of the intermediate StateManagerInput, or to remove that middle man).

Looking at the next patch in this series, I think that it's using a worker thread to "push in" status updates -- which means that there are potentially multiple threads which are all accessing some shared state in the StateManagerInput. That shared state is not a std::atomic and it is not being locked. (There's no warning from TSAN because CMake says that there are no tests.)

State changes -- whenever there's a change of state in any source, the code in manager queries all of its inputs for an update ("all of these inputs" are effectively some wrapper caches, so it is not an expensive operation). I think that you can simply move that state into a map within the manager. What about something like this (pseudocode):

 enum State ...;
 class Manager;

 class Source {
   Source(std::shared_ptr<Manager>) : mgr(mgr) { mgr->plug(this, Unknown); }
   ~Source() { mgr->unplug(this); }
 protected:
   void update(State value) {
     mgr->update(this, value);
   }
 private:
   std::shared_ptr<Manager> mgr;
 };

 class Manager {
 public:
   Manager(std::function<void(State)> notifier) : notifier(notifier) {}

   void plug(Source* src, State state) {
     if map.contains(src) throw ...
     map[src] = state;
     recalc();
   }

   void unplug(Source* src) {
     if !map.contains(src) throw ...
     map.remove(src);
     recalc();
   }

   void update(Source* src, State state) {
     if !map.contains(src) throw ...
     map[src] = state;
     recalc();
   }

 private:
   void recalc() {
     auto res = find_max_element_in(states);
     notifier(res);
   }

   std::map<Source*, State> states;
   std::function<void(State)> notifier;
 };

 class DBusWhateverSource : public Source {
   Whatever(std::shared_ptr<Manager> mgr, string service, string interface, string blabla) {
     dbus::onchangeof(service, interface, blabla, [this](dbus::variant newValue) {
       // FIXME: compute state from that new value
       update(newValue);
     });
   }
 };

 void main() {
   auto mgr = std::make_shared<Manager>([](State s) {
     switch(s) {
     case Red:
       redLed->on();
       greenLed->off();
       break;
     case Green:
       greenLed->on();
       redLed->off();
       break;
     }
   });
   auto claSysrepo = std::make_shared<WhateverSource>(mgr, blah, blah);
   auto systemdSource = std::make_shared<AnotherDBusSource>(mgr, blah, blah);
   dbus::evntloop::run();
 }

For this to support multithreading, a very simple locking will be needed in Manager::update(), Manager::plug() and Manager::unplug(). But perhaps everything can just run in a single thread as long as you put all of your sources into one "DBus context" or whatever it is called in our dbus library.

Patch-set: 8
Reviewer: Gerrit User 1000000 <1000000@e32f8df8-3db0-4f76-a6c6-fe9d6e69dd68>
Label: Code-Review=-1
Work-in-progress: false