Update patch set 1
Patch Set 1:

> Or perhaps I can let sdbus-c++ build systemd manage that [1] in dependencies repo?

I would prefer not to do it that way. Only velia needs this dependency, and I hope that both velia and lldp-whatever :) projects are going to become "done" soon. That's why I would prefer to move that systemd dependency towards the leaf projects and not bundle that via submodules/dependencies.

Patch-set: 1