Use systemd 246 when building on CI
In the next patch, we are going to switch to Fedora 32 builds on CI.
However, Fedora 32 uses systemd 245.8 which unfortunately suffers from
the bug which affects function calls in sdbus-c++ [1].

Therefore we should use systemd<245 or systemd>245 which do not have the
bug.
CMake will fail with FATAL_ERROR if systemd 245 is found.

[1] https://github.com/Kistler-Group/sdbus-cpp/issues/106

Change-Id: I2499b43e5bd3c2a1d99bacb3179eb863c0f2bbf3