log order when unit state changes
While browsing velia's logs I found the order of events logged a little misleading.
When systemd unit changes state and input to StateManager changes, we
first log the input change and then the info about unit state change.

This patch logs the events in the correct order.

Before, the log entries when a systemd unit changed state looked like
this (timestamps stripped):

 [main] [debug] Input 0x15efe0c changed state to State::ERROR
 [main] [info] Output changed to State::ERROR
 [main] [trace] Systemd unit 'systemd-journal-upload.service' changed state (activating auto-restart)

With this fix it will look like this. It should me more accurate:

 [main] [trace] Systemd unit 'systemd-journal-upload.service' changed state (activating auto-restart)
 [main] [debug] Input 0x15efe0c changed state to State::ERROR
 [main] [info] Output changed to State::ERROR

Change-Id: I73fe87c28d0d564046598080f425ab6abeefce71