Ignore changes in selected systemd units
This commit implements a possibility of ignoring changes in specified
systemd unit as proposed by [1]. This can be useful in cases where
some of the units are behaving badly (but not affecting the health
of a whole system).
Add a CLI option `--systemd-ignore-unit` to let user specify which units
should be ignored.

[1]: https://github.com/CESNET/velia/issues/1

Change-Id: I5902151b4ac95bdcd9aacf29c0189fb62537497f