rauc: Install bundle
Implement installation of a new bundle via RAUC D-Bus API [1].

The D-Bus interface is asynchronous, i.e., after the InstallBundle
method is invoked, it returns immediately and status of the installation
is provided to the caller (or anyone) via property changes (see [2]).
After the installation is completed, a D-Bus signal Completed is
invoked.

In order to capture these notifications (property changes and signals)
we introduce three callback functions that needs to be passed to the
RAUC constructor. These callbacks notify the user of this class in the
following situations:
- operation changes (RAUC can perform only one operation at once, so the
  user can see when the RAUC is busy or idle),
- installation progresses, and
- installation is completed.

[1] https://rauc.readthedocs.io/en/v1.4/using.html#using-the-d-bus-api
[2] https://rauc.readthedocs.io/en/v1.4/using.html#installing-a-bundle

Change-Id: Id8114edc3a271b2fa40b7bcb1bd46b776c3584e1