firewall: Add sysrepo-nft support
The SysrepoFirewall class is used as translator between libyang trees
from sysrepo and a configuration "file" (it's a string) which is then
supplied to a consumer. This consumer enables me to test the
translation.

The YANG model is pretty complicated because of all the deviations from
original model. That's why the translation algorithm traverses all of
the tree so that no nodes are missed.

Change-Id: Ic6733787dc73aa6e0c25297785bed724c3d79e70