system: Rename Sysrepo to IETFSystem
In the next few commits we will start implementing czechlight-system
model, which is our custom model for (so far) firmware management.
This commit renames the original Sysrepo class to IETFSystem so it
better expresses that it support only ietf-system model.
The czechlight-system model will be handled by a separate class.

Also, move the IETFSystem class to a separate sysrepo namespace.

Change-Id: Ibc7dc6270424229dad3ad7c951d0e61a65a91ef0