utils: push values to sysrepo
Add a helper for setting Sysrepo values into a specified datastore. The
values are passed as a std::map (mapping XPath to a value) which
might simplify our life in future data manipulation.

There are two instances of the helper. The first one just pushes the data
into the current datastore of the session. The second one wraps around
the first one and it changes the session's datastore to the specified
one and restores the previous datastore after the changes are applied.

Change-Id: Ic0c374737047ef5944dbc4d8af34a6e364946c05