daemons: Move sdbus++ event loop handling
We are going to introduce multiple daemons in the velia project soon.
All the daemons will (so far) talk/listen to D-Bus using sdbus++ [1]
library. This library also support a basic event loop which we are
using and the parts that initialize the event loop are going to be
common among those daemons.

Therefore this patch factors out the event loop initialization and
operations into a separate header so that future daemons can use it too
without too much code duplication.

Also, originally the event loop std::thread object was constructed
inside the main() function. However, to simplify things, this patch
constructs the thread object as a global variable. We are using other
global variables anyway (sdbus++ D-Bus connection object), so we can
declare both these variables at once and the event loop code becomes
less scattered around the file.

[1] https://github.com/Kistler-Group/sdbus-cpp/

Change-Id: I9de64a9cbf8ddb347ef0a65efdce26ac3a230004