split Factory class
In the next few patches we are going to split veliad into velia and
velia-hardware. However, the current veliad uses a Factory class which
has two functions, one for velia(-health) and one for velia-hardware.

It is only logical to split this Factory class into two, one for the
health and one for the hardware part.
There is no point in the velia-hardware to depend on stuff from
velia-health and vice versa.

Change-Id: If935c21b5da98cf56713479635d6ae6266c680a7