daemons: Add daemonize option without sdbus++
Current daemons use sdbus++ library to start an event loop and keep
going until SIGTERM against the process is issued.
This is fine so far, however, in the next patch we introduce
veliad-hardware daemon, which (so far) does not use sdbus++ and therefore
it does not depend on the library.

We need the veliad-hardware daemon just to keep itself running, all the
Sysrepo background stuff is happening in the Sysrepo threads.

This is the same use case we encountered in the
lldp-systemd-networkd-sysrepo project and this patch introduces the same
code that achieves that. The code is copied from the commit
52f2fe693d20c6da803df6a64f29d0a209b75454.

Change-Id: I76ef6bf119361b0c6b3f763acb6bce4f8c620a2f