veliad-hardware: sysrepo daemon managing hardware
This patch factors out the Sysrepo implementation of ietf-hardware
model. We agreed to split velia into smaller parts and this is a logical
individual unit.
The original velia will implement only the health tracking.

In the future, veliad-hardware might start notifying the original veliad
daemon when something goes wrong in the hardware. This will probably be
done using some kind of IPC, like D-Bus.

Change-Id: Ib831b1fd97d8ec87a711321461a3d6a73b63bcf7