veliad: rename to veliad-health
As we are splitting the original veliad into smaller parts, the original
veliad is now only tracking health of the services. In the future, when
we connect the veliad-* processes with some IPC, it will also track the
health of the whole box itself (hardware warnings, optical components
warnings, etc.) and it will manipulate with the LEDs accordingly.

So, in order to better express the functionality of this daemon, the
original veliad is now called veliad-health.

Change-Id: I269b970360e804f05fb770e71b15de13c1301b9a