yang: provide firmware data via czechlight-system
This commit introduces a new revision (2021-01-25) of czechlight-system
YANG model. The model is used for various system operations in
czechlight boxes.

This revision adds reporting of firmware slot status into firmware
container.
The firmware-status list contains (for each slot present) an information
about the slot name, firmware version installed in the slot, the
installation time and the current slot state. The state is usually either
inactive or booted.

Change-Id: I5a3555dfb4569a68208b27b58808e21c91222b59