system: Rename CzechlightSystem to Firmware
By the time it happened that the czechlight-system YANG model is going
to support authentication and other system-related things. It would be
too much for a single Sysrepo class, so we are creating new ones. To
reflect this change, rename the CzechlightSystem class to Firmware,
because it will manage only the firmware-related stuff.

Change-Id: I49a344d98d42379fd7b231761dd99e8cb726f1a9