health: do not connect to sysrepo
This one does not require a sysrepo connection, so let's make sure it
does not even try to connect there.

I only noticed because this thing broke the boot on a CzechLight SDN
Coherent Add/Drop in the lab. On new sysrepo it is super-critical to
always start *after* the YANG bits are prepared. That is done in
br2-external in the `reset-sysrepo` package.

Change-Id: I9bfd53b5d4005bd185aee8690e379941c8195d73