system: Use two D-Bus RAUC proxy objects to avoid deadlock
This patch adds a second D-Bus proxy object for RAUC on a different
connection. This should avoid a deadlock happening in the sdbus-c++'s
code. Consider the following situation with two threads: T1 (executing
sdbus-c++ event loop) and T2 (executing Sysrepo callback) and a RAUC
installation invoked.

T1: Gets notified upon signal PropertyChanged on the property Progress.
    sdbus-c++ locks sdbus mutex for the current connection (so only one
    object can talk to sdbus at one point).
T1: Invokes callback that creates Sysrepo notification.
T1: Sysrepo starts validating the operational datastore and invokes
    pull operational data callback.
T2: Starts processing the operational data callback.
T2: Invokes a code that calls GetSlotStatus on RAUC via D-Bus
    (something that is done in the next commit) and gets stuck in
    sdbus-c++ code because it is trying to acquire a mutex currently
    held by T1.

In order not to get stuck in this situation we create two D-Bus proxy
objects. Both on different connections.
The first proxy is supposed to handle all the D-Bus signals. The second
proxy handles the method calls. Because both proxies are on different
connections, they do not share the same sdbus mutex in sdbus-c++ and
therefore the scenario described above won't result in a deadlock.

This could also be solved by not calling the GetSlotStatus method in the
pull ops data callback. However, we want those data to be provided by
Sysrepo and we want them to be "fresh". This would probably require some
kind of timer that would refresh the data every N time units and push
them into Sysrepo. But I don't want to mess velia project with timers now.

Change-Id: I209f57da247bc1745261c258f4e2db73f72a4f1a