tests: Don't use D-Bus custom names when mocking systemd's Manager
Because org.freedesktop.systemd1 busname is often used even as session
busname we had to improvise in systemd tests and we used custom name
cz.cesnet.systemd1 along with custom object paths and interfaces
when mocking systemd's D-Bus behaviour.
This is not necessary.
There is no need to create custom paths and interfaces, we can use the
standard ones. We only have to use another bus name.

Change-Id: Ied5dbfe9ad2fec2bd30787d5d6916bf5a5924fd9