system: Configure eth1 interface (on running system)
This commit is a part of implementation of network configuration for eth1
interface through Sysrepo and velia-system.

Implement /czechlight-system:networking/eth1-ip presence container in
running datastore. React to any change to the eth1-ip container in
running datastore.

If the container is missing, revert to "old" default configuration of
eth1, i.e., add it to the br0 bridge together with other interfaces. If
the container is present and static-ip-address leaf is set, remove eth1
from the br0 bridge and configure it with the static ip. If the
container is present and the leaf is not set, let eth1 obtain its IP via
DHCP.

The network is managed by systemd-networkd thus we implement the changes
via systemd-networkd network files [1]. The files can be placed in /etc,
/run and /usr directories (and they take precedence in this order). We
place the new network file in the /run directory and reload network
configuration via networkctl reload.

There is a slight catch in removing the interface from a bridge.
Apparently, systemd-networkd does not remove an interface from a bridge
if the new configuration does not contain Bridge settings [2].
However, reloading the new configuration and removing the bridge
manually using ip command works.

[1] https://www.freedesktop.org/software/systemd/man/systemd.network.html
[2] https://github.com/systemd/systemd/issues/8190

Change-Id: I8940985a88903c17ea61e55aced67b17fba4656a