system: Configure eth1 interface (on running system)
This commit is a part of implementation of network configuration for eth1
interface through Sysrepo and velia-system. It Implements
/czechlight-system:networking/standalone-eth1 presence container in
running datastore.

If the presence container is missing, the network configuration for eth1
interface resets to the "old" default configuration of eth1, i.e., add
it to the br0 bridge.
If the container is present, the interface is removed from the bridge and
its IP address is obtained via DHCP.

The network is managed by systemd-networkd thus we implement the changes
via systemd-networkd network files [1]. The files can be placed in /etc,
/run and /usr directories (and they take precedence in this order). We
place the new network file in the /run directory and reload network
configuration via networkctl reload.

There is a slight catch in removing the interface from a bridge.
Apparently, systemd-networkd does not remove an interface from a bridge
if the new configuration does not contain Bridge settings [2].
However, brinding the interface down and then reloading the new
configuration apparently works.

[1] https://www.freedesktop.org/software/systemd/man/systemd.network.html
[2] https://github.com/systemd/systemd/issues/8190

Change-Id: I8940985a88903c17ea61e55aced67b17fba4656a