system: Configure eth1 interface (persist the configuration)
This commit is a part of implementation of network configuration for eth1
interface through Sysrepo and velia-system.

On a change to czechlight-system:networking/standalone-eth1 presence
container in *startup* datastore, persist the configuration (systemd .network
file [1]) into /cfg tree. On the next boot, this configuration will be copied
into /run by a service in czechlight-cfg-fs package from br2-external project.

[1] https://www.freedesktop.org/software/systemd/man/systemd.network.html

Change-Id: I75933bd333c795527eab1470cb89108b2240dad0