system: Clear old install message in firmware updates
This patch implements clearing the old install message if new
installation is started. This could be confusing because it is not clear
to which installation the message refers.

Consider the scenario with two consecutive installations via RAUC. The
first installation fails and an error message is set into
/czechlight-system:firmware/installation/message leaf. Then another
installation could be started and while
/czechlight-system:firmware/installation/status correctly contains value
"in-progress", the message leaf still contains the old error message.

Change-Id: I5bcc12bd7a700aa846ea9f725e7c0e52880c5ec1