Reduce logging a bit
Now that there's no --log-level flag, a lot of these daemons are using
just two loggers. In case of the "hardware" service, it's the "sysrepo"
one and the "hardware" one, for example. Do not attempt to log
debug/trace messages via "main", that's useless as they are silently
discarded.

The critical exceptions are still logged via "main", that's OK because
of the default log level.

Also remove some log messages which are really not that informative. I
can imagine they were useful when writing this, but let's hope that we
won't be debugging failures such as "not being able to connect to
D-Bus", etc. These logs were now getting silently discarded anyway.

Change-Id: Ic6e7a8f79d1f2d51992b36c9e53fb568a8727e25
Fixes: 3da1467 ("Get rid of the --log-level option")