ietf-system: handle "contact" and "location"
These are part of the ietf-system ML, so, presumably, it makes some
sense to allow their modification. In the new sysrepo architecture, this
modification *is* supported out-of-box because they are part of a module
that is implemented. However, unless there's an active callback which
subscribes to their modifications, they won't appear in the operational
CB. That's confusing -- and it's confusing also in the `netconf-cli`
because its `get` currently retrieves data from the operational DS.

Change-Id: I4bd6f184c5780845c29ca9f1b4777e3d1d7248fe