yang: czechlight-network (ietf-interfaces, ietf-ip) model
Introduce ietf-interfaces (RFC 8343) [1] and ietf-ip (RFC 8344) [2]
models to velia.
We are going to implement these models for interface management.
However, in the first stages we don't want to implement interface
statistics (/ietf-interfaces:interfaces/interface/statistics) container.
Also, for simplicity, we want to start with read-only access for the
interface nodes. For that we introduce czechlight-network YANG model
which adds the mentioned limitations to the ietf-interface tree.

ietf-ip and iana-if-type YANG model files are taken from YangModels
repository [3].

[1] https://tools.ietf.org/html/rfc8343
[2] https://tools.ietf.org/html/rfc8344
[3] https://github.com/YangModels/yang/tree/fb9c099/standard/ietf/RFC

Change-Id: Ic9e03f298de818eaf3508f7397263e0453be6e2e