system: Don't reconfigure network interface if not necessary
It is not necessary to reload network if the new setting is the same as
the current one because the network outage may be long (locally it is
around 6 seconds).

This issue can happen on system startup.

1. cfg-restore-systemd-networkd service copies a configuration from /cfg
to /run/systemd/network directory.
2. velia-system starts up and try to reconfigure the network by writing
a .network file to the same directory. The content of the file is
generated from the setting in the running datastore.

However, this setting *should* be the same as the setting copied from /cfg
and therefore there is no point in reconfiguring the network.
It'd be different only if startup datastore and data in /cfg are, for some
reason, not synchronized (e.g. someone removes the file from /cfg).

Change-Id: I83b2f8fbcc61962ab7b264d591f546b4a3119590