system: Get link info from rtnetlink
This commit introduces Rtnetlink class. The purpose of the class is to
communicate with kernel using netlink protocol and provide a method that
allows to iterate over links so that the caller doesn't have to worry
about netlink initialization and/or cleanup and focus only on processing
the rtnl_link struct to obtain the desired data.

The netlink communication is implemented by Netlink Protocol Library
Suite (libnl) [1]. Its part (libnl-route) also provides some handy
shortcuts when dealing with NETLINK_ROUTE scokets.

[1] https://www.infradead.org/~tgr/libnl/

Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/br2-external/+/3978
Depends-on: https://cesnet-gerrit-public/c/CzechLight/br2-external/+/3889
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/br2-external/+/3889
Change-Id: I703c69827bb6b974d37b05a18f2af468685ec758