system: Work around RAUC not providing some properties
We encountered an issue where RAUC did not provide bundle.version and
installed.timestamp properties. It turns out that it was because of
missing /slot.raucs file.
However we expected that RAUC always provide all the required
properties and this issue then led to map::at throwing an exception.

This patch implements 'more defensive behaviour' when working with the
contents returned by RAUC by checking whether the properties are present
in the map.

Change-Id: Iefed40c6cef410103fe583bd4dbd08dc39f069da