utils: Allow to remove values from sysrepo
Allow removing nodes in valuesPush so we can change and/or remove some
values from Sysrepo in one go.

The deletion of a node is implemented as NETCONF (RFC 6241 [1]) "remove"
operation. The remove operation deletes the data if present. If not, the
operation is a no-op.
If "delete" operation was used, the data would be removed (if present)
too. But if the data are not present it would result in an error [1].

We'd like to avoid the error without implementing a function that checks
if the data are present (see next commits).

[1] https://tools.ietf.org/html/rfc6241#page-38

Change-Id: Id4a0a83265f36a7bdac30a1c2d8247d25551d598