tests: wait for firmware installation finished notification
The sysrepo_system-firmware.cpp test file is used for testing the
installation of a new firmware (via RAUC) and it also tests that all the
data are correctly passed to sysrepo.

However, we rarely see CI failures in this test that look like this
one:

 2021-03-08 19:45:52.065462 | w | /home/ci/src/cesnet-gerrit-public/CzechLight/velia/tests/sysrepo_system-firmware.cpp:163: FATAL ERROR: REQUIRE( dataFromSysrepo(client, "/czechlight-system:firmware/installation", SR_DS_OPERATIONAL) == expectedFinished ) is NOT correct!
 2021-03-08 19:45:52.065478 | w |   values: REQUIRE( {
 2021-03-08 19:45:52.065487 | w |   "/message": "",
 2021-03-08 19:45:52.065496 | w |   "/status": "in-progress",
 2021-03-08 19:45:52.065505 | w | } == {
 2021-03-08 19:45:52.065514 | w |   "/message": "",
 2021-03-08 19:45:52.065522 | w |   "/status": "succeeded",
 2021-03-08 19:45:52.065531 | w | } )

In other words, we expect the firmware install to be finished (and
succeed) however, sysrepo still reports "in-progress.
Yes, we wait for a limited amount of time (2s) while the installation
process should take at most +- 750ms. This should have been enough.
But probably it sometimes isn't.

This patch removes the 2s wait for an installation to complete. Instead,
we listen for sysrepo notifications that are sent throughout the
installation to report progress. After the last notification is
obtained, the installation should be complete.

Of course, we add the callback to EventWatcher class. The callback fires
after any notification received.

Change-Id: I3be2cce374304194752cfa0306bd6195cc3310b1