HIT FIX: go IPv6-only when eth1 is outside of the bridge
I'm getting some unexpected behavior on the L2 layer in the lab when
this is enabled (such as two hosts trying to use the exact same IPv4
address at the same time). I have no clue if it's due to a broken
el-cheapo switch or if there's some misconfiguration in the network, but
we cannot afford to loose connectivity now (and one of the conflicting
addresses belonged to the VPN gateway). Force IPv6-only for now.

Change-Id: I968ee45ee8454bce5f9911bcb32f1f91f78d8aad