network: check that the startup config file gets generated
Up until the recent sysrepo update, this file won't get created if the
data are already present in the startup DS at the time of the
subscription because SR_SUBSCR_ENABLED apparently has no effect with
SR_DS_STARTUP.

This is important, for several reasons:

- We've change semantics of the `standalone-eth1` stanza once already,
and if that file is not regenerated, we would run with an older version.
- There was a bug in the RAUC hook which would accidentally skip copying
of the `/cfg/network/` config file on FW reinstall.

As a result, the system would accidentally run with a wrong network
config momentarily, which is bad.

Now that this has been fixed in sysrepo, capture this behavior in a unit
test.

Change-Id: I90cc6fbbba018b03a77d1dc982bc21d7e79a1d2b
Bug: https://github.com/sysrepo/sysrepo/issues/2376
Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/dependencies/+/4086
Depends-on: https://cesnet-gerrit-public/c/CzechLight/dependencies/+/4086
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/4086