network: generate the startup config file harder
This file won't get created if the data are already present in the
startup DS at the time of the subscription because SR_SUBSCR_ENABLED
apparently has no effect with SR_DS_STARTUP. Fix that by always
executing that CB manually.

This is important, for several reasons:

- We've change semantics of the `standaalone-eth1` stanza once already,
and if that file is not regenerated, we would run with an older version.
- There was a bug in the RAUC hook which would accidentally skip copying
of the `/cfg/network/` config file on FW reinstall.

As a result, the system would accidentally run with a wrong network
config momentarily, which is bad.

Change-Id: I90cc6fbbba018b03a77d1dc982bc21d7e79a1d2b
Bug: https://github.com/sysrepo/sysrepo/issues/2376