yang: Introduce ietf-routing module with dependencies
We are also going to implement a model dealing with network routing
tables. Such model (ietf-routing) is published in in RFC 8349 [1].
It needs to be augmented for ipv4/ipv6 routing. All the models are
downloaded from YangModels GitHub repository [2] (commit
8dd16482bb74c6add0b187a0e7755f5549da1ccb).

Also note that ietf-routing (as well as ietf-interfaces) also model
*-state containers which are supposed to be `config false;`. We are
actually using config false too but on the configuration containers.
This is done via augmentation in czechlight-network model.

This, however, led to unpleasant situation where I have not realized
that ietf-routing:routing-state depends on
ietf-interfaces:interface-state and the model could not be installed
into sysrepo (see https://github.com/sysrepo/sysrepo/issues/2384).
The augmentation in czechlight-network model had to be altered as well.

[1] https://tools.ietf.org/html/rfc8349
[2] https://github.com/YangModels/yang/tree/8dd16482bb74c6add0b187a0e7755f5549da1ccb/standard/ietf/RFC

Change-Id: I7b7d51b5c67e34acb3fa05c3745335b0ee4957d0