system: Perform initial change callbacks from libnl later
libnl's managed cache can monitor changes in specified caches. However,
it doesn't fire any events in the first population of the cache. We
acknowledged that and fired the events manually in the Rtnetlink
constructor.
However, if the handler for the event calls some Rtnetlink methods back,
the instance is not yet fully constructed and segfaults happen.

Change-Id: I80e2536cfae1444b242900918ee6647f0489f6f1