tests: Basic test for ietf-interfaces and routing
The test operates on a dummy interface that is created. Only privileged
user can create the interface using iproute2 and therefore we use sudo
to call the iproute2 binary (I was too lazy to use rtnetlink messages
for operating with the links and routes).

Because of that the test is "hidden" from the basic test suite. You
shall invoke cmake with -D=TEST_NETWORK_WITH_SUDO=ON to compile the test binary.

Also bump CMake minimum required version: find_program(... REQUIRED) is
supported since version 3.18 (and the minimum supported version
requirement is not mentioned until the 3.20 docs [1]). However, this is
`sudo` and `sudo` is magic, typically not installed as world-readable,
just world-executable (or magic-group-executable), and that means that
cmake's find_program ignores that. That is, unless CMP0109 [2] is set to
`NEW`, and that one requires CMake 3.19+.

Finally, bcause /usr/sbin/sudo is not in a non-privileged user's $PATH,
adjust the CI environment.

[1] https://cmake.org/cmake/help/v3.20/command/find_program.html
[2] https://cmake.org/cmake/help/latest/policy/CMP0109.html

Change-Id: Ia3dad61b6baff3b4d6e55a333f1375984bb937e5