tests: Basic test for ietf-interfaces and routing
The test operates on a dummy interface that is created. Only privileged
user can create the interface using iproute2 and therefore we use sudo
to call the iproute2 binary (I was too lazy to use rtnetlink messages
for operating with the links and routes).

Because of that the test is "hidden" from the basic test suite. You
shall invoke cmake with -DTESTS_ROOT=ON to compile the test binary.

Change-Id: Ia3dad61b6baff3b4d6e55a333f1375984bb937e5