yang: Add read-only LED status to czechlight-system
This change models the state of LEDs. It adds a container for LED
"stuff" and a list of LEDs. Each LED has a current and maximum
brightness. According to kernel source [1] and LED sysfs docs [2],
both these values can be unsigned ints at most.

Also bump revision of the czechlight-system YANG model to 2021-04-19.

[1] https://github.com/torvalds/linux/blob/3fb6d0e00efc958d01c2f109c8453033a2d96796/include/linux/leds.h#L64
[2] https://github.com/torvalds/linux/blob/edef2fa29739172643dbe9f44fb38d23793816d6/Documentation/ABI/testing/sysfs-class-led

Change-Id: I6bceef4b5d258d136af1515553e85f2f1fa6cbfa