health: compute status LED brightness from maximum brightness
We used hardcoded value 255 for maximum brightness but LEDs on Clearfog
devices have 256 maximum brightness.  This led to situations like that
/czechlight-system:leds reports only 99% brightness for status LEDs
(and that is not pretty -- why only 99?).

To deal with this we read maximum brightness in LED Sysfs driver
constructor and use that when setting the actual color of the status
LED.

Change-Id: I6c2b6b2d31a0744da3ad089b341473d6e2613ff1