utils: add valuesPush/valuesToYang overloads with data stored in std::vector
We have encountered an issue [1] with key-less lists in libyang where
passing the data to utils::valuesPush and utils::valuesToYang with
map<string, string> was causing problems because of the default
comparator on keys (strings). It did not worked as expected with
numerical values inside the XPaths (e.g., /a:b/lst[10] is sorted before
/a:b/lst[1]).

Let's introduce overloads for utils::valuesPush and utils::valuesToYang
which work with the data sorted by us by using std::vector instead of
std::map.

See next commit for the usage.

[1] https://github.com/CESNET/velia/issues/2

Change-Id: Idec62cab1ccb3b3d30a50d563f20fcb08121c208