system: Correctly fill YANG list in routing
We fill the YANG tree from a mapping XPath -> value which is implemented
by a map<string, string>. This data structure is by default sorted by
the key (string) and in this sorted order it is added to the YANG tree.

There is a problem that when there are at least 10 values in the list
(i.e., we have indices 1..10) then XPaths
 /ietf-routing:routing/ribs/rib/routes/route[1]
and
 /ietf-routing:routing/ribs/rib/routes/route[10]
are in the wrong order. The element indexed by 10 is sorted before the
element indexed with 1. This makes the "10" element to be inserted
first.
However, as the list has no keys, the "10" element will be processed by
libyang as the first element of the list effectively making it first.
And this causes a problem, because next insertion (element "1") can't be
inserted at position one, because (as the libyang says)
"Node already exists".

Let's work around this by allowing the data to be inserted in the
libyang tree to be sorted manually by storing them in the std::vector.

Bug: https://github.com/CESNET/velia/issues/2
Change-Id: If2992a801ade65a8c25f693c49916d458f3e26c5